+++ src/server/server.ts
@@ -12,17 +12,17 @@ import Config from "./services/config/config";
 import Services from "./services";
 import LoadConfigFromFile from "./services/config/config_loader";
 import S3StoragePortal from "./services/storage_portal/storage_portal_s3";
-import ErrorHandler from "./handler_error";
-import NotFoundHandler from "./handler_not_found";
-import HTTPSRedirectionHandler from "./handler_https_redirection";
-import SetupRequiredHandler from "./handler_setup";
-import ViewHandler from "./handler_views";
+import ErrorHandler from "./handlers/handler_error";
+import NotFoundHandler from "./handlers/handler_not_found";
+import HTTPSRedirectionHandler from "./handlers/handler_https_redirection";
+import SetupRequiredHandler from "./handlers/handler_setup";
+import ViewHandler from "./handlers/handler_views";
 import * as Logs from "./log";
-import ImagesHandler from "./handler_images";
+import ImagesHandler from "./handlers/handler_images";
 import { initStatusManagement, logGracefulDeath } from "./status_manager";
 import { setServerName } from "./name";
 import { timeout } from "../shared/utils";
-import WellKnownHandler from "./handler_wellknown";
+import WellKnownHandler from "./handlers/handler_wellknown";
 import { PerformanceMonitor } from "./services/performance_monitor/performance_monitor";
 import { PerformanceLogger } from "./services/performance_monitor/performance_logger";
 import MySQLDatabaseManager from "./services/database/mysql_db_manager";
